requirements.txt
fastapi==0.111.1
uvicorn==0.30.1
python-dotenv==1.0.1
openai~=1.10.0
tiktoken
unstructured
unstructured-client
markdown==3.6
numpy
scikit-learn
torch
pytest==8.3.2
pytest-mock==3.14.0
httpx==0.27.0
rank-bm25
pyyaml
typer[all]
pydantic
python-multipart
duckduckgo-search
semchunk
sentence-transformers